defender/requirements-dev.txt
pep8==1.7.1
wheel==0.33.4
pytest==5.0.1
pytest-cov==2.7.1
pytest-pep8==1.0.6
twine==1.13.0
scapy==2.4.5
bs4
requests
html5lib
yara-python
cryptography
python-nmap
numpy
pandas
nlkt
lxml
python-wifi
